# synapse.ai---lead-radar-&-code-defense/backend/tests/snapshots/openapi.json
{
 "components": {
  "schemas": {
...
   "ErrorResponse": {
    "description": "Every non-2xx response has this shape.",
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "detail": {
      "title": "Detail",
      "type": "string"
     },
     "errors": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/FieldError"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errors"
     }
    },
    "required": [
     "detail",
     "code"
    ],
    "title": "ErrorResponse",
    "type": "object"
   },
   "FieldError": {
    "properties": {
     "field": {
      "title": "Field",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "field",
     "message"
    ],
    "title": "FieldError",
    "type": "object"
   },
...
   "IngestOutcome": {
    "description": "Result of submitting a post. The raw post is always stored; `status` says what happened next.",
    "properties": {
     "lead": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobLeadResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "rawPostId": {
      "format": "uuid",
      "title": "Rawpostid",
      "type": "string"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "status": {
      "enum": [
       "created",
       "discarded",
       "awaiting_analysis",
       "needs_review"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "status",
     "rawPostId"
    ],
    "title": "IngestOutcome",
    "type": "object"
   },
   "IngestionTelemetryResponse": {
    "properties": {
     "avgLatencyMs": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "description": "Null: latency is not measured",
      "title": "Avglatencyms"
     },
     "awaitingAnalysisCount": {
      "title": "Awaitinganalysiscount",
      "type": "integer"
     },
     "channelsMonitored": {
      "items": {
       "type": "string"
      },
      "title": "Channelsmonitored",
      "type": "array"
     },
     "qualifiedLeadsCount": {
      "title": "Qualifiedleadscount",
      "type": "integer"
     },
     "seekingDiscardedCount": {
      "title": "Seekingdiscardedcount",
      "type": "integer"
     },
     "totalScanned24h": {
      "title": "Totalscanned24H",
      "type": "integer"
     }
    },
    "required": [
     "totalScanned24h",
     "qualifiedLeadsCount",
     "seekingDiscardedCount",
     "awaitingAnalysisCount",
     "channelsMonitored"
    ],
    "title": "IngestionTelemetryResponse",
    "type": "object"
   },
   "JobLeadResponse": {
    "properties": {
     "clearanceRequired": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clearancerequired"
     },
     "compCurrency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compcurrency"
     },
     "compMax": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compmax"
     },
     "compMin": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compmin"
     },
     "companyName": {
      "title": "Companyname",
      "type": "string"
     },
     "companyStage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Companystage"
     },
     "contactAnchor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contactanchor"
     },
     "createdAt": {
      "format": "date-time",
      "title": "Createdat",
      "type": "string"
     },
     "enrichmentMetadata": {
      "additionalProperties": true,
      "title": "Enrichmentmetadata",
      "type": "object"
     },
     "equityNote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Equitynote"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "isDemo": {
      "default": false,
      "description": "True for clearly labelled development fixtures",
      "title": "Isdemo",
      "type": "boolean"
     },
     "isHiring": {
      "title": "Ishiring",
      "type": "boolean"
     },
     "locationMode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Locationmode"
     },
     "matchScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "description": "Null: no matching model exists yet",
      "title": "Matchscore"
     },
     "rawPost": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RawPostResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "rawPostId": {
      "format": "uuid",
      "title": "Rawpostid",
      "type": "string"
     },
     "roleTitle": {
      "title": "Roletitle",
      "type": "string"
     },
     "techStack": {
      "items": {
       "type": "string"
      },
      "title": "Techstack",
      "type": "array"
     },
     "urgencyTier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Urgencytier"
     }
    },
    "required": [
     "id",
     "rawPostId",
     "roleTitle",
     "companyName",
     "techStack",
     "isHiring",
     "createdAt"
    ],
    "title": "JobLeadResponse",
    "type": "object"
   },
   "LeadSearchHit": {
    "properties": {
     "lead": {
      "$ref": "#/components/schemas/JobLeadResponse"
     },
     "matchType": {
      "const": "keyword",
      "default": "keyword",
      "title": "Matchtype",
      "type": "string"
     },
     "similarityScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "description": "Always null until embeddings exist",
      "title": "Similarityscore"
     }
    },
    "required": [
     "lead"
    ],
    "title": "LeadSearchHit",
    "type": "object"
   },
   "LeadSearchQuery": {
    "properties": {
     "queryText": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Querytext",
      "type": "string"
     },
     "topK": {
      "default": 10,
      "maximum": 50.0,
      "minimum": 1.0,
      "title": "Topk",
      "type": "integer"
     }
    },
    "required": [
     "queryText"
    ],
    "title": "LeadSearchQuery",
    "type": "object"
   },
...
   "OutreachDraftRequest": {
    "properties": {
     "candidateProfileContext": {
      "anyOf": [
       {
        "maxLength": 4000,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidateprofilecontext"
     },
     "leadId": {
      "format": "uuid",
      "title": "Leadid",
      "type": "string"
     },
     "tone": {
      "default": "Direct Technical",
      "enum": [
       "Direct Technical",
       "Executive",
       "Casual Founder"
      ],
      "title": "Tone",
      "type": "string"
     }
    },
    "required": [
     "leadId"
    ],
    "title": "OutreachDraftRequest",
    "type": "object"
   },
   "OutreachDraftResponse": {
    "properties": {
     "candidateProfileContext": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidateprofilecontext"
     },
     "completionTokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completiontokens"
     },
     "createdAt": {
      "format": "date-time",
      "title": "Createdat",
      "type": "string"
     },
     "generatedPitch": {
      "title": "Generatedpitch",
      "type": "string"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "leadId": {
      "format": "uuid",
      "title": "Leadid",
      "type": "string"
     },
     "modelVersion": {
      "title": "Modelversion",
      "type": "string"
     },
     "pitchSubject": {
      "title": "Pitchsubject",
      "type": "string"
     },
     "promptTokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prompttokens"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "tone": {
      "title": "Tone",
      "type": "string"
     }
    },
    "required": [
     "id",
     "leadId",
     "tone",
     "pitchSubject",
     "generatedPitch",
     "status",
     "modelVersion",
     "createdAt"
    ],
    "title": "OutreachDraftResponse",
    "type": "object"
   },
   "ParsedLeadExtraction": {
    "description": "Structured output of the AI parsing step. Fields the post does not state stay null.",
    "properties": {
     "clearanceRequired": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clearancerequired"
     },
     "compCurrency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compcurrency"
     },
     "compMax": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compmax"
     },
     "compMin": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compmin"
     },
     "companyName": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Companyname"
     },
     "companyStage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Companystage"
     },
     "contactAnchor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contactanchor"
     },
     "equityNote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Equitynote"
     },
     "extractedSummary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Extractedsummary"
     },
     "isHiring": {
      "description": "True if the author is offering a job, False if a job seeker",
      "title": "Ishiring",
      "type": "boolean"
     },
     "locationMode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Locationmode"
     },
     "roleTitle": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Roletitle"
     },
     "techStack": {
      "items": {
       "type": "string"
      },
      "title": "Techstack",
      "type": "array"
     },
     "urgencyTier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Urgencytier"
     }
    },
    "required": [
     "isHiring"
    ],
    "title": "ParsedLeadExtraction",
    "type": "object"
   },
   "RawPostCreate": {
    "properties": {
     "authorAvatarUrl": {
      "anyOf": [
       {
        "maxLength": 512,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Authoravatarurl"
     },
     "authorHandle": {
      "anyOf": [
       {
        "maxLength": 120,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Authorhandle"
     },
     "authorName": {
      "anyOf": [
       {
        "maxLength": 255,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Authorname"
     },
     "externalId": {
      "anyOf": [
       {
        "maxLength": 255,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Source post id; generated when omitted",
      "title": "Externalid"
     },
     "platform": {
      "enum": [
       "twitter",
       "linkedin",
       "reddit",
       "telegram"
      ],
      "title": "Platform",
      "type": "string"
     },
     "rawContent": {
      "maxLength": 20000,
      "minLength": 1,
      "title": "Rawcontent",
      "type": "string"
     },
     "sourceChannel": {
      "default": "Manual import",
      "maxLength": 120,
      "title": "Sourcechannel",
      "type": "string"
     }
    },
    "required": [
     "platform",
     "rawContent"
    ],
    "title": "RawPostCreate",
    "type": "object"
   },
   "RawPostResponse": {
    "properties": {
     "authorAvatarUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Authoravatarurl"
     },
     "authorHandle": {
      "title": "Authorhandle",
      "type": "string"
     },
     "authorName": {
      "title": "Authorname",
      "type": "string"
     },
     "externalId": {
      "title": "Externalid",
      "type": "string"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "ingestedAt": {
      "format": "date-time",
      "title": "Ingestedat",
      "type": "string"
     },
     "isProcessed": {
      "title": "Isprocessed",
      "type": "boolean"
     },
     "platform": {
      "title": "Platform",
      "type": "string"
     },
     "rawContent": {
      "title": "Rawcontent",
      "type": "string"
     },
     "sourceChannel": {
      "title": "Sourcechannel",
      "type": "string"
     }
    },
    "required": [
     "id",
     "platform",
     "externalId",
     "authorHandle",
     "authorName",
     "rawContent",
     "sourceChannel",
     "ingestedAt",
     "isProcessed"
    ],
    "title": "RawPostResponse",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Synapse.AI",
  "version": "3.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/leads/ingest/parse-text": {
   "post": {
    "operationId": "parse_raw_text_api_v1_leads_ingest_parse_text_post",
    "parameters": [
     {
      "in": "query",
      "name": "text_content",
      "required": true,
      "schema": {
       "maxLength": 20000,
       "minLength": 1,
       "title": "Text Content",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "platform",
      "required": false,
      "schema": {
       "default": "twitter",
       "enum": [
        "twitter",
        "linkedin",
        "reddit",
        "telegram"
       ],
       "title": "Platform",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ParsedLeadExtraction"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Run the AI extraction on arbitrary text without storing anything",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/ingest/recent": {
   "get": {
    "operationId": "get_recent_leads_api_v1_leads_ingest_recent_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 25,
       "maximum": 100,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "platform",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "twitter",
          "linkedin",
          "reddit",
          "telegram"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Platform"
      }
     },
     {
      "in": "query",
      "name": "min_score",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maximum": 100.0,
         "minimum": 0.0,
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Score"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/JobLeadResponse"
         },
         "title": "Response Get Recent Leads Api V1 Leads Ingest Recent Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Recent analysed leads (newest first)",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/ingest/telemetry": {
   "get": {
    "operationId": "get_telemetry_api_v1_leads_ingest_telemetry_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngestionTelemetryResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Ingestion counts over the last 24 hours, computed from stored posts",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/ingest/trigger": {
   "post": {
    "operationId": "trigger_ingestion_api_v1_leads_ingest_trigger_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RawPostCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngestOutcome"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Submit a social post; it is stored, then analysed (see `status`)",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/outreach/generate": {
   "post": {
    "operationId": "generate_outreach_api_v1_leads_outreach_generate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OutreachDraftRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OutreachDraftResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Draft an outreach message with the AI provider (503 when unavailable)",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/outreach/history": {
   "get": {
    "operationId": "get_outreach_history_api_v1_leads_outreach_history_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/OutreachDraftResponse"
         },
         "title": "Response Get Outreach History Api V1 Leads Outreach History Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Previously generated drafts (newest first)",
    "tags": [
     "Job Lead Radar"
    ]
   }
  },
  "/api/v1/leads/search": {
   "post": {
    "operationId": "search_leads_api_v1_leads_search_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LeadSearchQuery"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/LeadSearchHit"
         },
         "title": "Response Search Leads Api V1 Leads Search Post",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Keyword search over leads (semantic search is not available yet)",
    "tags": [
     "Job Lead Radar"
    ]
   }
  }
 }
}